The Multivariate Continuous Phase-Type Distribution

The Multivariate Continuous Phase-Type Distribution

dMPH(x, obj)

qMPH(p, obj)

pMPH(q, obj)

rMPH(n, obj)

Arguments

x, q
vector of quantiles.

obj
an object of class mult_cont_phase_type.

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dMPH gives the density, pMPH gives the distribution function, qMPH gives the quantile function, and rMPH generates random deviates.

The length of the result is determined by n for rMPH, and is the maximum of the lengths of the numerical arguments for the other functions.

The numerical arguments other than n are recycled to the length of the result. Only the first elements of the logical arguments are used.

Details

If the object provided is multivariate, each row of the result will corresponds to each univariate reward transformation. For dMPH, qMPH and pMPH, the inputs x, p and q can be matrices where in row i the i_th reward transformation and in col j the j_th value of x, p or q tested.

Functions

dMPH: Density function for the multivariate continuous phase-type distribution.

qMPH: Quantile function for the multivariate continuous phase-type distribution.

pMPH: Distribution function for the multivariate continuous phase-type distribution.

rMPH: Random number generator for the multivariate continuous phase-type distribution.

See also

Examples

cont_phase_type <- matrix(c(-3, 0, 1,
                            2, -3, 1,
                            1, 1, -2), ncol = 3)
R <- matrix(c(0, 1, 1,  2,
                 2, 1, 5,  2,
                 0, 1, 10, 2), nrow = 3, ncol=4, byrow=TRUE)
Y <- MPH(cont_phase_type, reward_mat = R)

#> Warning: 
#>  The initial probability vector is automatically generated.

dMPH(3:4, Y)

#>            [,1]       [,2]
#> [1,] 0.04978707 0.01831564
#> [2,] 0.08879596 0.04274991
#> [3,] 0.08694436 0.06568935
#> [4,] 0.13639222 0.09291062

pMPH(1.45, Y)

#>           [,1]
#> [1,] 0.7654297
#> [2,] 0.6202987
#> [3,] 0.2238680
#> [4,] 0.3415354

qMPH(0.5, Y)

#>           [,1]
#> [1,] 0.6931545
#> [2,] 1.0850215
#> [3,] 4.3387820
#> [4,] 2.1700430

rMPH(6, Y)

#>            [,1]      [,2]      [,3]     [,4]      [,5]      [,6]
#> [1,]  0.4508105 2.1002810 1.3664337 0.824683 2.3952692 0.7687702
#> [2,]  0.4913370 0.4934892 0.4989442 1.643679 0.8364322 3.0821149
#> [3,] 16.0316888 2.3824020 3.2356199 3.895310 9.3804166 3.8848225
#> [4,]  6.3383525 1.4363330 3.5830396 3.951306 1.5188263 2.7974890