Rate matrix and state space of the block counting process

RateMatAndStateSpace finds the state space and corresponding rate matrix for the block counting process for a number of samples n in the standard coalescent.

RateMAndStateSpace(n)

Arguments

n
Number of samples

Value

List consisting of RateM: Rate matrix StSpM: Matrix with rows corresponding to the states A state is a n-dimensional row vector (a1,...,an). For example the beginning state is (n,0,...,0), the next state is (n-2,1,0,...,0), and the ending state is (0,...,0,1)

Examples

RateMAndStateSpace(8)

#> $RateM
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]  -28   28    0    0    0    0    0    0    0     0     0     0     0
#>  [2,]    0  -21    6   15    0    0    0    0    0     0     0     0     0
#>  [3,]    0    0  -15    0    5   10    0    0    0     0     0     0     0
#>  [4,]    0    0    0  -15    1    8    6    0    0     0     0     0     0
#>  [5,]    0    0    0    0  -10    0    0    4    6     0     0     0     0
#>  [6,]    0    0    0    0    0  -10    0    1    3     3     3     0     0
#>  [7,]    0    0    0    0    0    0  -10    0    3     0     6     1     0
#>  [8,]    0    0    0    0    0    0    0   -6    0     0     0     0     3
#>  [9,]    0    0    0    0    0    0    0    0   -6     0     0     0     1
#> [10,]    0    0    0    0    0    0    0    0    0    -6     0     0     1
#> [11,]    0    0    0    0    0    0    0    0    0     0    -6     0     0
#> [12,]    0    0    0    0    0    0    0    0    0     0     0    -6     0
#> [13,]    0    0    0    0    0    0    0    0    0     0     0     0    -3
#> [14,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22]
#>  [1,]     0     0     0     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0
#>  [8,]     3     0     0     0     0     0     0     0     0
#>  [9,]     2     2     1     0     0     0     0     0     0
#> [10,]     0     4     0     1     0     0     0     0     0
#> [11,]     2     1     1     2     0     0     0     0     0
#> [12,]     0     0     6     0     0     0     0     0     0
#> [13,]     0     0     0     0     2     1     0     0     0
#> [14,]    -3     0     0     0     1     1     1     0     0
#> [15,]     0    -3     0     0     1     0     1     1     0
#> [16,]     0     0    -3     0     0     2     0     1     0
#> [17,]     0     0     0    -3     0     1     2     0     0
#> [18,]     0     0     0     0    -1     0     0     0     1
#> [19,]     0     0     0     0     0    -1     0     0     1
#> [20,]     0     0     0     0     0     0    -1     0     1
#> [21,]     0     0     0     0     0     0     0    -1     1
#> [22,]     0     0     0     0     0     0     0     0     0
#> 
#> $StSpM
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#>  [1,]    8    0    0    0    0    0    0    0
#>  [2,]    6    1    0    0    0    0    0    0
#>  [3,]    5    0    1    0    0    0    0    0
#>  [4,]    4    2    0    0    0    0    0    0
#>  [5,]    4    0    0    1    0    0    0    0
#>  [6,]    3    1    1    0    0    0    0    0
#>  [7,]    2    3    0    0    0    0    0    0
#>  [8,]    3    0    0    0    1    0    0    0
#>  [9,]    2    1    0    1    0    0    0    0
#> [10,]    2    0    2    0    0    0    0    0
#> [11,]    1    2    1    0    0    0    0    0
#> [12,]    0    4    0    0    0    0    0    0
#> [13,]    2    0    0    0    0    1    0    0
#> [14,]    1    1    0    0    1    0    0    0
#> [15,]    1    0    1    1    0    0    0    0
#> [16,]    0    2    0    1    0    0    0    0
#> [17,]    0    1    2    0    0    0    0    0
#> [18,]    1    0    0    0    0    0    1    0
#> [19,]    0    1    0    0    0    1    0    0
#> [20,]    0    0    1    0    1    0    0    0
#> [21,]    0    0    0    2    0    0    0    0
#> [22,]    0    0    0    0    0    0    0    1
#>